fixed various UI issues
- fixed double click issue closes #56
- fixed spacing issue around list titles
- fixed issue with rm anova factors control
