+++ client/analysisui/main.less
@@ -11,6 +11,7 @@
 @import (inline) "gridcombobox.css";
 @import (inline) "rmanovafactorscontrol.css";
 @import (inline) "gridvariablestargetlist.css";
+@import (inline) "gridtargetlist.css";
 
 
 * {
